+++ lib/config/environment/environment.dart
@@ -6,7 +6,7 @@ import 'package:flutter_template/common/service/log_history/log_history_service_
 import 'package:flutter_template/config/app_config.dart';
 import 'package:flutter_template/config/environment/build_types.dart';
 import 'package:flutter_template/persistence/storage/config_storage/config_storage.dart';
-import 'package:flutter_template/util/log_history.dart';
+import 'package:flutter_template/common/utils/logger/log_history.dart';
 import 'package:logger/logger.dart';
 import 'package:surf_logger/surf_logger.dart' as surf;
