+++ lib/features/app/di/app_scope.dart
@@ -16,7 +16,7 @@ import 'package:flutter_template/config/environment/environment.dart';
 import 'package:flutter_template/features/navigation/service/router.dart';
 import 'package:flutter_template/persistence/storage/theme_storage/theme_storage.dart';
 import 'package:flutter_template/persistence/storage/theme_storage/theme_storage_impl.dart';
-import 'package:flutter_template/util/default_error_handler.dart';
+import 'package:flutter_template/common/utils/default_error_handler.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 /// Scope of dependencies which need through all app's life.
